Finished Heroes of Pymoli
Finished out remaining tasks, made sure the code is stable and consistent. Only issue is the last bit regarding the way formatting as currency seems to negatively affect sorting ability. Left the last column on the last two tables as a raw number instead of currency because it was messing up the sort function.